Add files via upload
This is a problem I met when taking a test online. I doubt if it has a simple mathematical solution as I was expected to solve it with
pencil and paper in no more than 8 minutes. The description of the problem is as follows:

Suppose there are 10 islands in a row, for simplicity, let's denote them by island_1, island_2,...island_10. There are 9 pairs of bridges between these islands and connects consecutive islands, that is, there is a pair of bridges that connects island_1 and island_2, another pair that connects island_2 to island_3 and so on.

For each pair of bridges, one is made of iron and the other is made of wood. When a person steps onto the wooden bridge, it immediately breaks and the person will fall into the water and the flow of ocean will take him/her back to the island_1.  If he/she walks on the iron bridge, nothing unfortunate will happen and he/she will arrive at the next island.

The question is, suppose every time the person needs to cross the bridge to go to th